opimize and reduce component reader time
