+++ src/interface/property.d.ts
@@ -1,3 +1,5 @@
+import { Effect, Reducer, Subscription } from 'umi';
+
 export interface ISinglePropertyType {
   id: number;
   address: string;
@@ -8,4 +10,23 @@ export interface ISinglePropertyType {
 export interface FromValues {
   address: string;
   valuation: number;
+}
+
+export interface PropertyState {
+  properties: ISinglePropertyType[];
+}
+
+export interface PropertyModelType {
+  namespace: 'property';
+  state: PropertyState;
+  reducers: {
+    propertyList: Reducer<PropertyState>;
+  };
+  effects: {
+    getAllList: Effect;
+    postProperty: Effect;
+  };
+  subscriptions: {
+    setup: Subscription;
+  };
 }
